+++ spec/Pim/Bundle/CatalogBundle/Manager/ProductManagerSpec.php
@@ -5,11 +5,7 @@ namespace spec\Pim\Bundle\CatalogBundle\Manager;
 use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Builder\ProductBuilder;
-use Pim\Bundle\CatalogBundle\Event\ProductEvents;
-use Pim\Bundle\CatalogBundle\Manager\MediaManager;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
-use Pim\Bundle\CatalogBundle\Model\AvailableAttributes;
-use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Repository\AssociationTypeRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\AttributeOptionRepositoryInterface;
@@ -32,7 +28,6 @@ class ProductManagerSpec extends ObjectBehavior
         ProductSaver $productSaver,
         ProductSaver $productBulkSaver,
         EventDispatcherInterface $eventDispatcher,
-        MediaManager $mediaManager,
         ProductBuilder $builder,
         ProductRepositoryInterface $productRepository,
         AssociationTypeRepositoryInterface $associationTypeRepository,
@@ -53,7 +48,6 @@ class ProductManagerSpec extends ObjectBehavior
             $productSaver,
             $productBulkSaver,
             $eventDispatcher,
-            $mediaManager,
             $builder,
             $productRepository,
             $associationTypeRepository,